requirements.txt
Pillow>=9.0.0
PyYAML>=6.0
